Renames domain classes for consitency
